Extract substitution helpers into yaml/substitution.py
Second PR in the helpers/yaml split arc (follows #796). Moves
parse_substitution_ref and rewrite_name_or_substitution
byte-for-byte from yaml/__init__.py into the new
yaml/substitution.py submodule, plus the _PURE_SUBSTITUTION_REF
regex they share.

substitution.py imports the scalar helpers it needs
(_safe_yaml_scalar, _strip_yaml_quotes, read_yaml_scalar,
rewrite_yaml_scalar) from the parent package via
"from . import ...". Those will move into scalar.py in PR 4;
the import path stays valid because __init__.py re-exports
the same names from scalar.py at that point.

The re-export goes at the BOTTOM of __init__.py (after the
scalar definitions) so substitution.py's package-attribute
imports resolve correctly when the submodule loads.
PEP 484 redundant-alias form so ruff/mypy treat them as
intentional re-exports — external callers
(devices/mutations_clone.py, tests) keep using
"from .helpers.yaml import parse_substitution_ref" unchanged.

__init__.py: 1077 -> 1015 lines (-62). substitution.py: 83 lines.

All 3391 non-e2e tests pass; ruff + mypy clean.
